refactor: merge into one Database
