+++ app/src/main/java/com/github/onlynotesswent/model/note/NoteViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import com.github.onlynotesswent.model.cache.NoteDatabase
+import com.github.onlynotesswent.model.cache.CacheDatabase
 import com.google.firebase.Firebase
 import com.google.firebase.firestore.firestore
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -50,7 +50,7 @@ class NoteViewModel(private val repository: NoteRepository) : ViewModel() {
       initializer {
         NoteViewModel(
             NoteRepositoryFirestore(
-                Firebase.firestore, NoteDatabase.getNoteDatabase(context), context))
+                Firebase.firestore, CacheDatabase.getDatabase(context), context))
       }
     }
   }
